efi: libstub: remove DT dependency from generic stub
Refactor the generic EFI stub entry code so that all the dependencies on
device tree are abstracted and hidden behind a generic efi_boot_kernel()
routine that can also be implemented in other ways. This allows users of
the generic stub to avoid using FDT for passing information to the core
kernel.

Change-Id: I46c8004ea076f7e50c5ffad09f46c176cd009745
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>